+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "altinn-app-frontend",
-  "version": "2.4.3",
+  "version": "2.5.0",
   "description": "",
   "main": "index.js",
   "scripts": {
