+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/features/form/layout/formLayoutSagas.ts
@@ -1,7 +1,7 @@
 import { SagaIterator } from 'redux-saga';
 import { fork } from 'redux-saga/effects';
-import { watchFetchFormLayoutSaga, watchFetchFormLayoutSettingsSaga, watchUpdateCurrentViewSaga } from './fetch/fetchFormLayoutSagas';
-import { watchUpdateFocusSaga, watchUpdateAutoSave, watchUpdateRepeatingGroupsSaga } from './update/updateFormLayoutSagas';
+import { watchFetchFormLayoutSaga, watchFetchFormLayoutSettingsSaga } from './fetch/fetchFormLayoutSagas';
+import { watchUpdateFocusSaga, watchUpdateAutoSave, watchUpdateRepeatingGroupsSaga, watchUpdateCurrentViewSaga } from './update/updateFormLayoutSagas';
 
 // eslint-disable-next-line func-names
 export default function* (): SagaIterator {
